Middleware implementation
There is repetition in validating user permissions, especially for checking if a user is authorized to update or delete a group.

So i created 2 middlewares, they check for the group_id if its in the request and if its valid then the first one checks if the authenticated user is the owner or not, and the second one check if the user belongs to the group.